app/requirements.txt
streamlit==1.45.0
mysql-connector-python==8.2.0
folium==0.15.1
streamlit-folium==0.18.0
python-dotenv
Pillow
opencv-python
numpy